% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean_names.R
\name{func_clean_names}
\alias{func_clean_names}
\title{Função criada para limpar nomes de banco de dados, padronizando}
\usage{
func_clean_names(df, debug)
}
\description{
Autor: Mikael
 Data: 29/11/22
}
