% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_colnames_df.R
\name{func_get_colnames_df}
\alias{func_get_colnames_df}
\title{Funcoes para receber nomes de variaveis (colunas em df, unique nas tabelas)
  
 Autor: Mikael
 Data: 29/11/22}
\usage{
func_get_colnames_df(df)
}
\description{
Funcoes para receber nomes de variaveis (colunas em df, unique nas tabelas)
  
 Autor: Mikael
 Data: 29/11/22
}
